Experiment with different zero crossing algorithm
methods in ZeroCrossingCounter.h and FrequencyCalculation.h added (usuall name *Alt). Instead of counting zero crossings across a buffer, it writes to buffer until 3 zero crossings are detected. 3 zero crossings = 1 cycle. That cycle length is stored to a pointer and the writeHeadPos returns to 0. The lengths in that pointer are averaged and converted to frequency. Seems more accurate.